添加距离过远boss回血的goal，修复树魔二次挑战的bug

+++ src/main/java/com/gaboj1/tcr/entity/custom/boss/yggdrasil/MagicProjectile.java
@@ -11,12 +11,12 @@ import net.minecraft.world.phys.EntityHitResult;
 import org.jetbrains.annotations.NotNull;
 
 public class MagicProjectile extends LlamaSpit {
-    private int damage = 0;
+    private float damage = 0;
     public MagicProjectile(EntityType<? extends MagicProjectile> p_37248_, Level p_37249_) {
         super(p_37248_, p_37249_);
     }
 
-    public void setDamage(int damage) {
+    public void setDamage(float damage) {
         this.damage = damage;
     }
